+++ projects/ngx-smooth-scroll/src/lib/@types.d.ts
@@ -3,6 +3,14 @@ export interface NgxSmoothScrollOption {
     timingFunction?: string;
     alignX?: 'start' | 'center' | 'end';
     alignY?: 'start' | 'center' | 'end';
+    stopOnArrival?: boolean;
+}
+
+export interface NgxSmoothScrollDirectiveOption extends NgxSmoothScrollOption {
+    wheel?: NgxSmoothScrollOption;
+    keydown?: NgxSmoothScrollOption;
+    mouse?: NgxSmoothScrollOption;
+    touch?: NgxSmoothScrollOption;
 }
 
 export interface Coords {
@@ -12,7 +20,10 @@ export interface Coords {
 
 export type bezierArray = [number, number, number, number]
 
-export type NgxSmoothScrollKeyCodeMap = Map<'normal' | 'reverse', number[]>;
+export interface NgxSmoothScrollKeyCode {
+    'forward'?: Array<number>;
+    'reverse'?: Array<number>;
+}
 
 export interface NgxSmoothScrollBeforeAnimateEvent {
     currentIndex: number;
@@ -23,4 +34,9 @@ export interface NgxSmoothScrollAfterAnimateEvent {
     prevIndex: number;
     currentIndex: number;
     scrollCoords: Coords;
+}
+
+export interface ListenerData {
+    [key: string]: any;
+    events: Array<() => void>;
 }