+++ projects/ngx-smooth-scroll/src/lib/ngx-smooth-scroll-keyboard.directive.spec.ts
@@ -1,8 +1,0 @@
-import { NgxSmoothScrollKeyboardDirective } from './ngx-smooth-scroll-keyboard.directive';
-
-describe('NgxSmoothScrollKeyboardDirective', () => {
-  it('should create an instance', () => {
-    const directive = new NgxSmoothScrollKeyboardDirective();
-    expect(directive).toBeTruthy();
-  });
-});