+++ projects/ngx-smooth-scroll/src/lib/ngx-smooth-scroll-wheel.directive.spec.ts
@@ -1,8 +1,0 @@
-import { NgxSmoothScrollWheelDirective } from './ngx-smooth-scroll-wheel.directive';
-
-describe('NgxSmoothScrollDirective', () => {
-  it('should create an instance', () => {
-    const directive = new NgxSmoothScrollWheelDirective();
-    expect(directive).toBeTruthy();
-  });
-});