+++ projects/ngx-smooth-scroll/src/lib/ngx-smooth-scroll.directive.spec.ts
@@ -1,0 +1,8 @@
+import { NgxSmoothScrollDirective } from './ngx-smooth-scroll.directive';
+
+describe('NgxSmoothScrollDirective', () => {
+  it('should create an instance', () => {
+    const directive = new NgxSmoothScrollDirective();
+    expect(directive).toBeTruthy();
+  });
+});