+++ projects/ngx-smooth-scroll/src/lib/ngx-smooth-scroll.service.ts
@@ -1,5 +1,5 @@
 /** Native Modules */
-import { Injectable } from '@angular/core';
+import { Injectable, ElementRef } from '@angular/core';
 import { Observable } from 'rxjs';
 
 /** Custom Modules */
@@ -14,19 +14,19 @@ import { Coords, NgxSmoothScrollOption } from './@types';
 })
 export class NgxSmoothScrollService {
 
-  public createInstance(containerEl: HTMLElement, childSelector?: string): NgxSmoothScroll {
+  public createInstance(containerEl: HTMLElement | ElementRef, childSelector?: string): NgxSmoothScroll {
     return new NgxSmoothScroll(containerEl, childSelector);
   }
 
-  public scrollTo(containerEl: HTMLElement, destination: Coords, options: NgxSmoothScrollOption): Observable<Coords> {
+  public scrollTo(containerEl: HTMLElement | ElementRef, destination: Coords, options: NgxSmoothScrollOption): Observable<Coords> {
     return this.createInstance(containerEl).scrollTo(destination, options);
   }
 
-  public scrollToElement(containerEl: HTMLElement, childEl: HTMLElement, options?: NgxSmoothScrollOption): Observable<Coords> {
+  public scrollToElement(containerEl: HTMLElement | ElementRef, childEl: HTMLElement | ElementRef, options?: NgxSmoothScrollOption): Observable<Coords> {
     return this.createInstance(containerEl).scrollToElement(childEl, options);
   }
 
-  public scrollToIndex(containerEl: HTMLElement, childSelector: string, index: number, options?: NgxSmoothScrollOption): Observable<Coords> {
+  public scrollToIndex(containerEl: HTMLElement | ElementRef, childSelector: string, index: number, options?: NgxSmoothScrollOption): Observable<Coords> {
     return this.createInstance(containerEl, childSelector).scrollToIndex(index, options);
   }