+++ projects/ngx-smooth-scroll/src/public-api.ts
@@ -5,6 +5,5 @@
 export { NgxSmoothScrollService } from './lib/ngx-smooth-scroll.service';
 export { NgxSmoothScrollModule } from './lib/ngx-smooth-scroll.module';
 export { NgxSmoothScroll } from './lib/core/smooth-scroll.core';
-export { NgxSmoothScrollWheelDirective } from './lib/ngx-smooth-scroll-wheel.directive';
-export { NgxSmoothScrollKeyboardDirective } from './lib/ngx-smooth-scroll-keyboard.directive';
-export { NgxSmoothScrollOption, Coords, NgxSmoothScrollKeyCodeMap, NgxSmoothScrollBeforeAnimateEvent, NgxSmoothScrollAfterAnimateEvent } from './lib/@types';
+export { NgxSmoothScrollDirective } from './lib/ngx-smooth-scroll.directive';
+export { NgxSmoothScrollOption, Coords, NgxSmoothScrollKeyCode, NgxSmoothScrollBeforeAnimateEvent, NgxSmoothScrollAfterAnimateEvent, NgxSmoothScrollDirectiveOption } from './lib/@types';